exclude staff from link removal

+++ monitors/removeLinks.js
@@ -1,10 +1,9 @@
-const {Monitor} = require('klasa');
+const { Monitor } = require('klasa');
 
 module.exports = class extends Monitor {
 
     constructor(...args) {
         super(...args, {
-            name: 'responder',
             enabled: true,
             ignoreBots: false,
             ignoreSelf: true,
@@ -16,23 +15,23 @@ module.exports = class extends Monitor {
         });
     }
 
-    run(message) {
+    run (message) {
         if (message.type !== 'DEFAULT')
             return null;
-        if (/(https?:\/\/[^\s]+)/g.test(message.content) && message.member.settings.xp < 128)
-        {
+        const { permission } = await this.client.permissionLevels.run(message, 4);
+        if (/(https?:\/\/[^\s]+)/g.test(message.content) && message.member.settings.xp < 128 && !permission) {
             message.channel.send(`:x: <@${message.author.id}>, You must be level 3 (128 XP) or above to post links in this guild.`)
-                    .then((msg) => {
-                        setTimeout(function () {
-                            msg.delete();
-                        }, 10000);
-                    });
+                .then((msg) => {
+                    setTimeout(function () {
+                        msg.delete();
+                    }, 10000);
+                });
             message.delete();
             return null;
         }
     }
 
-    async init() {
+    async init () {
     }
 
 };
